lesson_5 - rename lection into lecture - PR
