--- src/schoolOnline/lection/Lection.java
@@ -1,11 +1,0 @@
-package schoolOnline.lection;
-
-public class Lection {
-    public static int lectionCount;
-    private int id;
-    String name;
-    int quantity;
-    public Lection(){
-        lectionCount++;
-    }
-}
