--- src/schoolOnline/lection/LectionServe.java
@@ -1,7 +1,0 @@
-package schoolOnline.lection;
-
-public class LectionServe {
-    public static Lection lectionCreate(){
-        return new Lection();
-    }
-}
