--- src/schoolOnline/lection/additionalMaterial/AdditionalMaterialServe.java
@@ -1,4 +1,0 @@
-package schoolOnline.lection.additionalMaterial;
-
-public class AdditionalMaterialServe {
-}
