--- src/schoolOnline/lection/homework/HomeworkServe.java
@@ -1,4 +1,0 @@
-package schoolOnline.lection.homework;
-
-public class HomeworkServe {
-}
