+++ src/schoolOnline/lecture/Lecture.java
@@ -1,0 +1,11 @@
+package schoolOnline.lecture;
+
+public class Lecture {
+    public static int lectureCount;
+    private int id;
+    String name;
+    int quantity;
+    public Lecture(){
+        lectureCount++;
+    }
+}
