+++ src/schoolOnline/lecture/LectureServe.java
@@ -1,0 +1,7 @@
+package schoolOnline.lecture;
+
+public class LectureServe {
+    public static Lecture lectureCreate(){
+        return new Lecture();
+    }
+}
