+++ src/schoolOnline/lecture/additionalMaterial/AdditionalMaterialServe.java
@@ -1,0 +1,4 @@
+package schoolOnline.lecture.additionalMaterial;
+
+public class AdditionalMaterialServe {
+}
