+++ src/schoolOnline/lecture/homework/Homework.java
@@ -1,4 +1,4 @@
-package schoolOnline.lection.homework;
+package schoolOnline.lecture.homework;
 
 public class Homework {
     private int id;
