+++ src/schoolOnline/lecture/homework/HomeworkServe.java
@@ -1,0 +1,4 @@
+package schoolOnline.lecture.homework;
+
+public class HomeworkServe {
+}
